mgr/dashboard: Forms now wait for all data to load until displayed
Add directive to handle the display of the form.
It will display the form only when loading is finished.
Otherwise it will show a loading message or error message.

Add a new class that should be extended by all form components.
For now it only has methods for dealing with loading,
but this could be improved later.

Fixes: https://tracker.ceph.com/issues/44912

Signed-off-by: Tiago Melo <tmelo@suse.com>