crimson: distinguish record and block relative paddrs
Blocks get read independently of the surrounding record,
so paddr's embedded directly in a block need to refer
to other blocks within the same record by a block_relative
addr relative to the block's own offset.  By contrast,
deltas to existing blocks need to use record_relative
addrs relative to the first block of the record.

This patch distinguishes the two kinds of relative paddr
(mainly for debugging purposes) and adapts cache, journal,
etc to use the appropriate types.

Signed-off-by: Samuel Just <sjust@redhat.com>