qa/suites/rados/thrash-old-clients: disable verify_ceph_hash
in this test, older ceph clients are installed on el7, but the ceph
cluster is deployed using cephadm, which in turn pulls ceph container
images built using the ceph being tested on el8.

since we've dropped the build of master on el7, there is no need to
verify if ceph package is available if cephadm is used for deploying the
cluster.

Signed-off-by: Kefu Chai <kchai@redhat.com>