osd: enforce the CRUSH bucket peering constraints on peering
We enforce the rules in choose_acting(), which is pretty simple.

More complicatedly, we also extend calc_replicated_acting to try
and extend the acting set with appropriate choices when we would
otherwise block peering due to site constraints. It doesn't
strive to provide a minimal set but it does have some basic smarts.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>