os/bluestore: Track per zone metadata
For every object we maintain <zone_num+oid, offset> tuple in the key-value
store.  When a new object written to a zone, we insert the corresponding tuple
to the database.  When an object is truncated, we remove the corresponding
tuple.  When an object is overwritten, we remove the old tuple and insert a new
tuple corresponding to the new location of the object.  The cleaner can now
identify live objects within the zone <zone_num> by enumerating all the keys
starting with <zone_num> prefix.

Signed-off-by: Abutalib Aghayev <agayev@cs.cmu.edu>