rgw: rgw-orphan-list -- fix interaction, quoting, and percentage calc
The interactive mode wasn't working due to prompts going to stdout
instead of stderr. If a space should appear in temporary file, it will
generate a shell error, so quoting was added. Furthermore if there are
no objects found in a pool, a divide by zero error will be
generated. This commit addresses these issues.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>