crimson/os/seastore/cache: don't mark invalid extents clean
In the event that an extent is created and removed in the same
transaction (invalid extent in fresh_block_list), update
block specific metadata but don't add to cache and especially
do not mark clean.  LBAManager::complete_commit implementations
are meant to use skip based on state in the same way.

Signed-off-by: Samuel Just <sjust@redhat.com>