msg,messages: remove MMonMetadata
MMonMetadata is sent to pre-luminous monitors and handled only for
backward compatibility with pre-luminous monitors. and luminous is
the last release we were still sending this message. but there is no
need for pacific to be compatible with pre-luminous monitors because of
the +2 policy.

so let's drop this message from the source tree.

Signed-off-by: Kefu Chai <kchai@redhat.com>