msg/async/ProtocolV2: allow rxbuf/txbuf get bigger in testing
We have a kernel client test case that constructs huge auth tickets
to exercise the three related code paths in the kernel.  One of the
tickets is bigger than 1000000 bytes, as required for triggering the
third code path.

We haven't bumped into this assert earlier because the kernel client
is still on msgr v1.  However, "rbd map" and "rbd unmap" commands
started connecting to the cluster in commit bbe8c8555fd4 ("rbd: delay
determination of default pool name") and that happens via msgr v2.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>