cephadm: implement a webservice to offer host metadata to mgr
Gathering the data on a host is time consuming, so this patch give
cephadm a daemon mode that handles the data gathering tasks
itself, so mgr can just ping an API endpoint. This provides better
scale and improves the data currency to orchestrator.

This patch introduces a couple of new classes to cephadm to handle
daemon mode supporting by some compatibility patches to allow
it to integrate with existing function flows.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>