common/admin_socket: always validate the parameters
this change addresses the regression introduced by
dc75434ebbf0e30b9ac56b83c3712e00c4dd4d5d.

before this change, we reply the client with a failure only if
both the prefix and the arguments do not match.

after this change, both the prefix and argument are checked.

Fixes: https://tracker.ceph.com/issues/47179
Signed-off-by: Kefu Chai <kchai@redhat.com>