crimson/os/seastore/cache: dirty next in replace_extent always
replace_extent is always used to replace an extent with its next
version, so next is dirty whether or not prev was.

Doing this in complete_commit posed a problem as the newly dirtied
extent wouldn't behave as such until after commit.  Latter portions
try_construct_record dealing with retired extents assume that any
extents mutated and retired in the same transaction are already
dirty.

Signed-off-by: Samuel Just <sjust@redhat.com>