crimson/net: cleanup interfaces to start and stop a messenger
There is no on-going iterations in the new
ChainedDispatchers::ms_dispatch() implementation, so we no longer need
to worry about removing dispatchers when stopping the messenger. So the
""boost::intrusive::slist" is not needed, and we can use cleaner
interfaces to start and stop the messenger.

Also fixed an regression issue in perf_crimson_msgr caused by
ChainedDispatchers.

Signed-off-by: Yingxin Cheng <yingxin.cheng@intel.com>