cmake: stop rebuilding rocksdb everytime
this change was originally introduced as a part of
bb9139fdf777d1b759608ff7294b730668383cae, and latter migrated / changed
in the current form. but the idea is the same: to rebuild rocksdb even
if the stamp file shows that it has been built. there is no need to do
so, as we don't hack RocksDB as we used to. also,it is distracting to
check this log message when rebuilding the tree. so drop it.

Signed-off-by: Kefu Chai <kchai@redhat.com>